+++ tests/integration_tests/game_layer/observations/test_node_observations.py
@@ -39,6 +39,7 @@ def test_host_observation(simulation):
         folders=[],
         network_interfaces=[],
         file_system_requires_scan=True,
+        include_users=False,
     )
 
     assert host_obs.space["operating_status"] == spaces.Discrete(5)